Confusion Matrix for SVM
This is my initial confusion matrix for the SVM model before mitigating for overfitting. This indicates a false negative or type 2 error seems to occur more (identifying as human, when it is actually AI generated). I personally believe this error is better because I think the effects of classifying a human written essay as AI is worse than missing an AI written essay.